+++ all-is-cubes-desktop/src/main.rs
@@ -342,7 +342,9 @@ async fn create_universe(
                 .map_err(anyhow::Error::from)
         }
         UniverseSource::File(path) => {
-            all_is_cubes_port::load_universe_from_file(yield_progress, &*path).await
+            all_is_cubes_port::load_universe_from_file(yield_progress, &*path)
+                .await
+                .map_err(anyhow::Error::from)
         }
     }?;
     universe_progress_bar.finish();